+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [v0.3.4]
+- Added more details to some validation errors. [#91](https://github.com/xmidt-org/ancla/pull/91)
+- Fixed naming mismatch for Validator config: changed name in Config struct to what we send in yaml. [#91](https://github.com/xmidt-org/ancla/pull/91)
+- Bumped chrysom to v0.6.0: updated code to create basic and listener clients. [#91](https://github.com/xmidt-org/ancla/pull/91)
+
 ## [v0.3.3]
 - Removed legacy logic condition for inserting a webhook. [#85](https://github.com/xmidt-org/ancla/pull/85)
 - Changed webhook address handling. [#84](https://github.com/xmidt-org/ancla/pull/84)
@@ -78,7 +83,8 @@ internalWebhooks. [#80](https://github.com/xmidt-org/ancla/pull/80)
 ## [v0.1.0]
 - Initial release
 
-[Unreleased]: https://github.com/xmidt-org/ancla/compare/v0.3.3...HEAD
+[Unreleased]: https://github.com/xmidt-org/ancla/compare/v0.3.4...HEAD
+[v0.3.4]: https://github.com/xmidt-org/ancla/compare/0.3.3...v0.3.4
 [v0.3.3]: https://github.com/xmidt-org/ancla/compare/0.3.2...v0.3.3
 [v0.3.2]: https://github.com/xmidt-org/ancla/compare/0.3.1...v0.3.2
 [v0.3.1]: https://github.com/xmidt-org/ancla/compare/0.3.0...v0.3.1